dashboard/app: chop off auth so it can be reused in syz-hub
